pyproject.toml
[project]
name = "backend-to-do-list"
version = "0.1.0"
description = ""
authors = [
    {name = "Leonardo de Almeida",email = "leo70x7000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend_to_do_list", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
